+++ packages/ignition/next.config.mjs
@@ -1,15 +1,22 @@
 import createMDX from "@next/mdx";
+import { createCssVariablesTheme } from "shiki";
 import rehypeShiki from "@shikijs/rehype";
 import rehypeSlug from "rehype-slug-custom-id";
 import remarkFrontmatter from "remark-frontmatter";
 import remarkMdxFrontmatter from "remark-mdx-frontmatter";
-import { createContentCollectionPlugin } from "@rocketclimb/content-collections";
+import { shikiColorToClassTransform } from "@rocketclimb/code-block/shiki-tranform";
+import { IconsManifest } from "rocketicons/data";
+
+const theme = createCssVariablesTheme({
+  name: "css-variables",
+  variablePrefix: "--shiki-",
+  variableDefaults: {},
+  fontStyle: true
+});
 
 const shikiOptions = {
-  themes: {
-    light: "dracula",
-    dark: "dracula"
-  }
+  theme,
+  transformers: [shikiColorToClassTransform()]
 };
 const withMDX = createMDX({
   options: {
@@ -21,6 +28,8 @@ const withMDX = createMDX({
   }
 });
 
+const packagesToOptimize = IconsManifest.map(({ id }) => `rocketicons/${id}`);
+
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   pageExtensions: ["js", "jsx", "mdx", "ts", "tsx"],
@@ -31,45 +40,8 @@ const nextConfig = {
   },
 
   experimental: {
-    optimizePackageImports: [
-      "rocketicons/ai",
-      "rocketicons/bi",
-      "rocketicons/bs",
-      "rocketicons/cg",
-      "rocketicons/ci",
-      "rocketicons/di",
-      "rocketicons/fa",
-      "rocketicons/fa6",
-      "rocketicons/fc",
-      "rocketicons/fi",
-      "rocketicons/gi",
-      "rocketicons/go",
-      "rocketicons/gr",
-      "rocketicons/hi",
-      "rocketicons/hi2",
-      "rocketicons/im",
-      "rocketicons/io",
-      "rocketicons/io5",
-      "rocketicons/lia",
-      "rocketicons/lu",
-      "rocketicons/md",
-      "rocketicons/pi",
-      "rocketicons/rc",
-      "rocketicons/ri",
-      "rocketicons/rx",
-      "rocketicons/si",
-      "rocketicons/sl",
-      "rocketicons/tb",
-      "rocketicons/tfi",
-      "rocketicons/ti",
-      "rocketicons/vsc",
-      "rocketicons/wi"
-    ]
+    optimizePackageImports: packagesToOptimize
   }
 };
 
-const withContentCollections = createContentCollectionPlugin({
-  skipWatcher: true
-});
-
-export default withContentCollections(withMDX(nextConfig));
+export default withMDX(nextConfig);