+++ packages/ignition/src/app/components/documentation/section-content.tsx
@@ -1,7 +1,7 @@
 import { PropsWithChildrenAndClassName } from "@/types";
 
 const SectionContent = ({ className, children }: PropsWithChildrenAndClassName) => (
-  <div className={`w-full px-0.5 shrink-0 ${className}`}>{children}</div>
+  <section className={`w-full px-0.5 grow shrink-0 ${className ?? ""}`}>{children}</section>
 );
 
 export default SectionContent;