+++ packages/ignition/src/app/components/url-observer.tsx
@@ -2,7 +2,16 @@ import { Suspense, useEffect } from "react";
 import { usePathname, useSearchParams } from "next/navigation";
 
 type UrlObserverProps = {
-  onChanges: ({ lastPath, hash }: { lastPath: string; hash: string }) => void;
+  onChanges: ({
+    pathName,
+    lastPath,
+    hash
+  }: {
+    pathName: string;
+    lastPath: string;
+    hash: string;
+    targetAnchor: string;
+  }) => void;
 };
 
 const UrlChangesObserver = ({ onChanges }: UrlObserverProps) => {
@@ -14,10 +23,13 @@ const UrlChangesObserver = ({ onChanges }: UrlObserverProps) => {
     const target = document.getElementById(targetId); // I've never saw that on a react code...
 
     const lastPath = pathName.split("/").pop() ?? "";
-    const hash =
-      window.location.hash && target ? targetId : (pathName.split("/").pop() as string);
+    const hash = window.location.hash;
+    const targetAnchor = hash && target ? targetId : lastPath;
+    console.log(hash);
     onChanges({
+      pathName,
       lastPath,
+      targetAnchor,
       hash
     });
   }, [pathName, searchParams]);