+++ packages/ignition/src/app/types/index.ts
@@ -7,3 +7,4 @@ export * from "./props-with-children-and-lang-and-slug-param";
 export * from "./props-with-lang-and-slug-param";
 export * from "./props-with-children-and-class-name";
 export * from "./props-with-class-name";
+export * from "./props-with-class-name-and-lang";