+++ packages/ignition/src/app/types/props-with-class-name-and-lang.ts
@@ -1,0 +1,4 @@
+import { PropsWithLang } from "./props-with-lang";
+import { PropsWithClassName } from "./props-with-class-name";
+
+export type PropsWithClassNameAndLang = PropsWithLang & PropsWithClassName;