+++ packages/ignition/src/config/site.ts
@@ -2,6 +2,7 @@ import { serverEnv } from "@/env/server";
 import { AvailableLanguages } from "@/app/types";
 
 const baseUrl = serverEnv.NEXT_PUBLIC_APP_URL || "https://rocketicons.io";
+const playgroundUrl = serverEnv.NEXT_PUBLIC_PLAYGROUND_URL || "https://playcode.io/1870276";
 const env =
   (process.env.NEXT_PUBLIC_VERCEL_ENV as "production" | "preview" | "development" | "local") ||
   "local";
@@ -19,7 +20,8 @@ export const siteConfig = {
     componentGroups: ["getting-started"]
   },
   isLocal: env === "local",
-  env
+  env,
+  playgroundUrl
 };
 
 export type SiteConfig = typeof siteConfig;